refactor: Combine WorkspacesPopover and WorkspacesSwitch
